Envia un missatge a un xat de Telegram

Aquest repositori est configurat per enviar missatges al grup de Telegram de la Comunitat en catal\u00E0 d'OpenStreetMap amb el bot monitorOSM_bot.

Usage

envia_missatge_telegram(
  missatge = "Hi ha canvis a [monitorOSM](https://osm-catalan.github.io/monitorOSM).",
  parse_mode = "Markdown"
)

Arguments

missatge

Text del missatge.

parse_mode

El missatge s'interpreta com a Markdown o HTML.

On i amb quin bot s'envia el missatge es configura amb les variables d'entorn R_TELEGRAM_BOT_MONITOROSM_BOT i TME_OSMCAT_CHATID. Per usar la funci a github, cal configurar les accions perqu carreguin aquestes variables d'entorn a partir dels secrets del repositori de github.